Collect build log files
It's a bit complicated because we can only figure out the location of
the logfile once it exists and the corresponding symlink is created.
We can't use the symlink directly because it points to the docker
internal directory structure (I guess we could improve that by copying
the logfiles at the end of the buildstep to a stable location).