Define codebases
The previous setup would apply the branch to all codebases, but we have
to deal with them separately.
The new setup is supposed to only listen for changes on the kube
codebase, but then take the default branch (as defined in the Git step) for the rest of the
codebases.